refactor: move mac clipboard utils to gui/
