+++ src/gui/osutils/macutils/MacPasteboard.cpp
