+++ src/gui/osutils/macutils/MacPasteboard.h
